toposort int

+++ src/docks.h
@@ -34,7 +34,7 @@ class DOCKS {
         int vertexExpMask;
         int vertexExp_1;
         byte* edgeArray;
-        byte* topoSort;
+        int* topoSort;
         static map<char, int> alphabetMap;
         string ALPHABET;
     DOCKS (int argK) {
@@ -183,7 +183,7 @@ class DOCKS {
         hittingNumArray = new double[(int)edgeNum];
         used = new byte[vertexExp];
         finished = new byte[vertexExp];
-        topoSort = new byte[vertexExp];
+        topoSort = new int[vertexExp];
         D = new float*[l + 1];
         float* Dpool = new float[(l+1)* vertexExp];
         for(int i = 0; i < l+1; i++, Dpool += vertexExp) D[i] = Dpool;
